+++ src/events/SubscriptionRequestEvent.php
@@ -17,11 +17,8 @@ use yii\base\Event;
  */
 class SubscriptionRequestEvent extends Event
 {
-    // Properties
-    // =========================================================================
-
     /**
      * @var array The subscription parameters
      */
-    public $parameters;
+    public array $parameters;
 }