+++ src/services/Invoices.php
@@ -24,9 +24,6 @@ use yii\base\Exception;
  */
 class Invoices extends Component
 {
-    // Constants
-    // =========================================================================
-
     /**
      * @event SaveInvoiceEvent The event that is triggered when an invoice is saved.
      * You may set [[SaveInvoiceEvent::isValid]] to `false` to prevent the invoice from being saved
@@ -44,7 +41,7 @@ class Invoices extends Component
      * });
      * ```
      */
-    const EVENT_BEFORE_SAVE_INVOICE = 'beforeSaveInvoice';
+    public const EVENT_BEFORE_SAVE_INVOICE = 'beforeSaveInvoice';
 
     /**
      * @event SaveInvoiceEvent The event that is triggered when an invoice is saved.
@@ -62,10 +59,7 @@ class Invoices extends Component
      * });
      * ```
      */
-    const EVENT_SAVE_INVOICE = 'afterSaveInvoice';
-
-    // Public Methods
-    // =========================================================================
+    public const EVENT_SAVE_INVOICE = 'afterSaveInvoice';
 
     /**
      * Returns all invoices for a subscription by its id.
@@ -93,7 +87,7 @@ class Invoices extends Component
     /**
      * Returns all invoices for a user by its id.
      *
-     * @param int $subscriptionId The subscription id.
+     * @param int $userId The user ID.
      * @return Invoice[]
      */
     public function getUserInvoices(int $userId): array
@@ -119,7 +113,7 @@ class Invoices extends Component
      * @param string $reference
      * @return Invoice|null
      */
-    public function getInvoiceByReference(string $reference)
+    public function getInvoiceByReference(string $reference): ?Invoice
     {
         $invoiceRow = $this->_createInvoiceQuery()
             ->where(['reference' => $reference])
@@ -183,9 +177,6 @@ class Invoices extends Component
         return false;
     }
 
-    // Private methods
-    // =========================================================================
-
     /**
      * Returns a Query object prepped for retrieving invoices.
      *
@@ -202,5 +193,4 @@ class Invoices extends Component
             ])
             ->from(['{{%stripe_invoices}} invoices']);
     }
-
 }